requirements.txt
# Dev environment
autopep8

# App
fastapi
sqlalchemy
pydantic
passlib[bcrypt]
python-jose[cryptography]
uvicorn
psycopg2-binary
python-dotenv
